refactor(gcf-util): move GCFLogger tests to separate file
