+++ packages/gcf-utils/test/gcf-utils.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {GCFBootstrapper, GCFLogger} from '../src/gcf-utils';
+import {GCFBootstrapper} from '../src/gcf-utils';
 import {describe, beforeEach, afterEach, it} from 'mocha';
 import {GitHubAPI} from 'probot/lib/github';
 import {Options} from 'probot';
@@ -21,9 +21,6 @@ import sinon from 'sinon';
 import nock from 'nock';
 import assert from 'assert';
 import {v1} from '@google-cloud/secret-manager';
-import {ObjectWritableMock} from 'stream-mock';
-import pino from 'pino';
-import {validateLogs, LogLine} from './test-helpers';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const repos = require('../../test/fixtures/repos.json');
@@ -36,374 +33,311 @@ function nockRepoList() {
     .reply(200, repos);
 }
 
-describe('gcf-util', () => {
-  describe('GCFLogger', () => {
-    describe('get()', () => {
-      it('returns a new pino-based logger instance on first call', () => {
-        const logger = GCFLogger.get();
-        assert.equal(logger.constructor.name, 'Pino');
-      });
-      it('returns the same logger instance on two consecutive calls', () => {
-        const logger1 = GCFLogger.get();
-        const logger2 = GCFLogger.get();
-        assert.deepEqual(logger1, logger2);
+describe('GCFBootstrapper', () => {
+  describe('gcf', () => {
+    let handler: (
+      request: express.Request,
+      response: express.Response
+    ) => Promise<void>;
+
+    const response: express.Response = express.response;
+    const sendStub: sinon.SinonStub<[object?], express.Response> = sinon.stub(
+      response,
+      'send'
+    );
+    const sendStatusStub: sinon.SinonStub<
+      [number],
+      express.Response
+    > = sinon.stub(response, 'sendStatus');
+
+    let req: express.Request;
+
+    const spy: sinon.SinonStub = sinon.stub();
+    let configStub: sinon.SinonStub<[], Promise<Options>>;
+
+    let bootstrapper: GCFBootstrapper;
+
+    let enqueueTask: sinon.SinonStub;
+
+    beforeEach(async () => {
+      req = express.request;
+
+      bootstrapper = new GCFBootstrapper();
+      configStub = sinon
+        .stub(bootstrapper, 'getProbotConfig')
+        .resolves({id: 1234, secret: 'foo', webhookPath: 'bar'});
+
+      enqueueTask = sinon.stub(bootstrapper, 'enqueueTask');
+
+      handler = await bootstrapper.gcf(async app => {
+        app.auth = () =>
+          new Promise<GitHubAPI>(resolve => {
+            resolve(GitHubAPI());
+          });
+        app.on('issues', spy);
+        app.on('schedule.repository', spy);
+        app.on('err', sinon.stub().throws());
       });
     });
 
-    describe('logger instance', () => {
-      let writeStream: ObjectWritableMock;
-      let logger: pino.Logger;
+    afterEach(() => {
+      sendStub.reset();
+      sendStatusStub.reset();
+      spy.reset();
+      configStub.reset();
+      enqueueTask.reset();
+    });
 
-      function readLogsAsObjects(writeStream: ObjectWritableMock): LogLine[] {
-        try {
-          writeStream.end();
-          const lines: string[] = writeStream.data;
-          return lines.map(line => JSON.parse(line));
-        } catch (error) {
-          throw new Error(`Failed to read stream: ${error}`);
-        }
-      }
+    it('calls the event handler', async () => {
+      req.body = {
+        installation: {id: 1},
+      };
+      req.headers = {};
+      req.headers['x-github-event'] = 'issues';
+      req.headers['x-github-delivery'] = '123';
+      req.headers['x-cloudtasks-taskname'] = 'my-task';
 
-      function testAllLevels() {
-        const levels: {[index: string]: number} = {
-          trace: 10,
-          debug: 20,
-          info: 30,
-          metric: 30,
-          warn: 40,
-          error: 50,
-        };
+      await handler(req, response);
 
-        for (const level of Object.keys(levels)) {
-          it(`logs ${level} level string`, () => {
-            logger[level]('hello world');
-            const loggedLines: LogLine[] = readLogsAsObjects(writeStream);
-            validateLogs(loggedLines, 1, ['hello world'], [], levels[level]);
-          });
+      sinon.assert.calledOnce(configStub);
+      sinon.assert.notCalled(sendStatusStub);
+      sinon.assert.calledOnce(sendStub);
+      sinon.assert.calledOnce(spy);
+    });
 
-          it(`logs ${level} level json`, () => {
-            logger[level]({hello: 'world'});
-            const loggedLines: LogLine[] = readLogsAsObjects(writeStream);
-            validateLogs(loggedLines, 1, [], [{hello: 'world'}], levels[level]);
-          });
-        }
-      }
+    it('does nothing if there are missing headers', async () => {
+      req.body = {
+        installation: {id: 1},
+      };
+      req.headers = {};
 
-      beforeEach(() => {
-        writeStream = new ObjectWritableMock();
-        logger = GCFLogger['initLogger'](undefined, writeStream);
-      });
+      await handler(req, response);
 
-      testAllLevels();
+      sinon.assert.calledOnce(configStub);
+      sinon.assert.notCalled(spy);
+      sinon.assert.notCalled(sendStub);
+      sinon.assert.calledWith(sendStatusStub, 400);
+    });
+
+    it('returns 500 on errors', async () => {
+      req.body = {
+        installtion: {id: 1},
+      };
+      req.headers = {};
+      req.headers['x-github-event'] = 'err';
+      req.headers['x-github-delivery'] = '123';
+      req.headers['x-cloudtasks-taskname'] = 'my-task';
+
+      await handler(req, response);
+
+      sinon.assert.calledOnce(configStub);
+      sinon.assert.notCalled(spy);
+      sinon.assert.notCalled(sendStatusStub);
+      sinon.assert.called(sendStub);
+    });
+
+    it('ensures that task is enqueued when called by scheduler for one repo', async () => {
+      req.body = {
+        installtion: {id: 1},
+        repo: 'firstRepo',
+      };
+      req.headers = {};
+      req.headers['x-github-event'] = 'schedule.repository';
+      req.headers['x-github-delivery'] = '123';
+      req.headers['x-cloudtasks-taskname'] = '';
+
+      await handler(req, response);
+
+      sinon.assert.calledOnce(enqueueTask);
+    });
+
+    it('ensures that task is enqueued when called by scheduler for many repos', async () => {
+      req.body = {
+        installtion: {id: 1},
+      };
+      req.headers = {};
+      req.headers['x-github-event'] = 'schedule.repository';
+      req.headers['x-github-delivery'] = '123';
+      req.headers['x-cloudtasks-taskname'] = '';
+      nockRepoList();
+
+      await handler(req, response);
+
+      sinon.assert.calledTwice(enqueueTask);
+    });
+
+    it('ensures that task is enqueued when called by Github', async () => {
+      req.body = {
+        installtion: {id: 1},
+      };
+      req.headers = {};
+      req.headers['x-github-event'] = 'another.name';
+      req.headers['x-github-delivery'] = '123';
+      req.headers['x-cloudtasks-taskname'] = '';
+
+      await handler(req, response);
+
+      sinon.assert.calledOnce(enqueueTask);
     });
   });
 
-  describe('GCFBootstrapper', () => {
-    describe('gcf', () => {
-      let handler: (
-        request: express.Request,
-        response: express.Response
-      ) => Promise<void>;
+  describe('loadProbot', () => {
+    let bootstrapper: GCFBootstrapper;
+    let configStub: sinon.SinonStub<[], Promise<Options>>;
 
-      const response: express.Response = express.response;
-      const sendStub: sinon.SinonStub<[object?], express.Response> = sinon.stub(
-        response,
-        'send'
-      );
-      const sendStatusStub: sinon.SinonStub<
-        [number],
-        express.Response
-      > = sinon.stub(response, 'sendStatus');
-
-      let req: express.Request;
-
-      const spy: sinon.SinonStub = sinon.stub();
-      let configStub: sinon.SinonStub<[], Promise<Options>>;
-
-      let bootstrapper: GCFBootstrapper;
-
-      let enqueueTask: sinon.SinonStub;
-
-      beforeEach(async () => {
-        req = express.request;
-
-        bootstrapper = new GCFBootstrapper();
-        configStub = sinon
-          .stub(bootstrapper, 'getProbotConfig')
-          .resolves({id: 1234, secret: 'foo', webhookPath: 'bar'});
-
-        enqueueTask = sinon.stub(bootstrapper, 'enqueueTask');
-
-        handler = await bootstrapper.gcf(async app => {
-          app.auth = () =>
-            new Promise<GitHubAPI>(resolve => {
-              resolve(GitHubAPI());
-            });
-          app.on('issues', spy);
-          app.on('schedule.repository', spy);
-          app.on('err', sinon.stub().throws());
-        });
-      });
-
-      afterEach(() => {
-        sendStub.reset();
-        sendStatusStub.reset();
-        spy.reset();
-        configStub.reset();
-        enqueueTask.reset();
-      });
-
-      it('calls the event handler', async () => {
-        req.body = {
-          installation: {id: 1},
-        };
-        req.headers = {};
-        req.headers['x-github-event'] = 'issues';
-        req.headers['x-github-delivery'] = '123';
-        req.headers['x-cloudtasks-taskname'] = 'my-task';
-
-        await handler(req, response);
-
-        sinon.assert.calledOnce(configStub);
-        sinon.assert.notCalled(sendStatusStub);
-        sinon.assert.calledOnce(sendStub);
-        sinon.assert.calledOnce(spy);
-      });
-
-      it('does nothing if there are missing headers', async () => {
-        req.body = {
-          installation: {id: 1},
-        };
-        req.headers = {};
-
-        await handler(req, response);
-
-        sinon.assert.calledOnce(configStub);
-        sinon.assert.notCalled(spy);
-        sinon.assert.notCalled(sendStub);
-        sinon.assert.calledWith(sendStatusStub, 400);
-      });
-
-      it('returns 500 on errors', async () => {
-        req.body = {
-          installtion: {id: 1},
-        };
-        req.headers = {};
-        req.headers['x-github-event'] = 'err';
-        req.headers['x-github-delivery'] = '123';
-        req.headers['x-cloudtasks-taskname'] = 'my-task';
-
-        await handler(req, response);
-
-        sinon.assert.calledOnce(configStub);
-        sinon.assert.notCalled(spy);
-        sinon.assert.notCalled(sendStatusStub);
-        sinon.assert.called(sendStub);
-      });
-
-      it('ensures that task is enqueued when called by scheduler for one repo', async () => {
-        req.body = {
-          installtion: {id: 1},
-          repo: 'firstRepo',
-        };
-        req.headers = {};
-        req.headers['x-github-event'] = 'schedule.repository';
-        req.headers['x-github-delivery'] = '123';
-        req.headers['x-cloudtasks-taskname'] = '';
-
-        await handler(req, response);
-
-        sinon.assert.calledOnce(enqueueTask);
-      });
-
-      it('ensures that task is enqueued when called by scheduler for many repos', async () => {
-        req.body = {
-          installtion: {id: 1},
-        };
-        req.headers = {};
-        req.headers['x-github-event'] = 'schedule.repository';
-        req.headers['x-github-delivery'] = '123';
-        req.headers['x-cloudtasks-taskname'] = '';
-        nockRepoList();
-
-        await handler(req, response);
-
-        sinon.assert.calledTwice(enqueueTask);
-      });
-
-      it('ensures that task is enqueued when called by Github', async () => {
-        req.body = {
-          installtion: {id: 1},
-        };
-        req.headers = {};
-        req.headers['x-github-event'] = 'another.name';
-        req.headers['x-github-delivery'] = '123';
-        req.headers['x-cloudtasks-taskname'] = '';
-
-        await handler(req, response);
-
-        sinon.assert.calledOnce(enqueueTask);
-      });
+    beforeEach(() => {
+      bootstrapper = new GCFBootstrapper();
+      configStub = sinon
+        .stub(bootstrapper, 'getProbotConfig')
+        .resolves({id: 1234, secret: 'foo', webhookPath: 'bar'});
     });
 
-    describe('loadProbot', () => {
-      let bootstrapper: GCFBootstrapper;
-      let configStub: sinon.SinonStub<[], Promise<Options>>;
-
-      beforeEach(() => {
-        bootstrapper = new GCFBootstrapper();
-        configStub = sinon
-          .stub(bootstrapper, 'getProbotConfig')
-          .resolves({id: 1234, secret: 'foo', webhookPath: 'bar'});
-      });
-
-      afterEach(() => {
-        configStub.reset();
-      });
-
-      it('gets the config', async () => {
-        await bootstrapper.loadProbot(async () => {
-          // Do nothing
-        });
-        sinon.assert.calledOnce(configStub);
-      });
-
-      it('caches the probot if initialized', async () => {
-        await bootstrapper.loadProbot(async () => {
-          // Do nothing
-        });
-        sinon.assert.calledOnce(configStub);
-        await bootstrapper.loadProbot(async () => {
-          // Do nothing again
-        });
-        sinon.assert.calledOnce(configStub);
-      });
+    afterEach(() => {
+      configStub.reset();
     });
 
-    describe('getProbotConfig', () => {
-      let bootstrapper: GCFBootstrapper;
-      let secretClientStub: v1.SecretManagerServiceClient;
-      let secretsStub: sinon.SinonStub;
-      let secretVersionNameStub: sinon.SinonStub;
-
-      beforeEach(() => {
-        secretClientStub = new v1.SecretManagerServiceClient();
-        bootstrapper = new GCFBootstrapper(secretClientStub);
-
-        secretVersionNameStub = sinon
-          .stub(bootstrapper, 'getLatestSecretVersionName')
-          .callsFake(() => {
-            return 'foobar';
-          });
+    it('gets the config', async () => {
+      await bootstrapper.loadProbot(async () => {
+        // Do nothing
       });
+      sinon.assert.calledOnce(configStub);
+    });
 
-      afterEach(() => {
-        secretsStub.reset();
-        secretVersionNameStub.reset();
+    it('caches the probot if initialized', async () => {
+      await bootstrapper.loadProbot(async () => {
+        // Do nothing
       });
+      sinon.assert.calledOnce(configStub);
+      await bootstrapper.loadProbot(async () => {
+        // Do nothing again
+      });
+      sinon.assert.calledOnce(configStub);
+    });
+  });
 
-      it('gets the config', async () => {
-        secretsStub = sinon
-          .stub(secretClientStub, 'accessSecretVersion')
-          .resolves([
-            {
-              payload: {
-                data: JSON.stringify({
-                  id: 1234,
-                  secret: 'foo',
-                  webhookPath: 'bar',
-                }),
-              },
+  describe('getProbotConfig', () => {
+    let bootstrapper: GCFBootstrapper;
+    let secretClientStub: v1.SecretManagerServiceClient;
+    let secretsStub: sinon.SinonStub;
+    let secretVersionNameStub: sinon.SinonStub;
+
+    beforeEach(() => {
+      secretClientStub = new v1.SecretManagerServiceClient();
+      bootstrapper = new GCFBootstrapper(secretClientStub);
+
+      secretVersionNameStub = sinon
+        .stub(bootstrapper, 'getLatestSecretVersionName')
+        .callsFake(() => {
+          return 'foobar';
+        });
+    });
+
+    afterEach(() => {
+      secretsStub.reset();
+      secretVersionNameStub.reset();
+    });
+
+    it('gets the config', async () => {
+      secretsStub = sinon
+        .stub(secretClientStub, 'accessSecretVersion')
+        .resolves([
+          {
+            payload: {
+              data: JSON.stringify({
+                id: 1234,
+                secret: 'foo',
+                webhookPath: 'bar',
+              }),
             },
-          ]);
-        await bootstrapper.getProbotConfig();
-        sinon.assert.calledOnce(secretsStub);
-        sinon.assert.calledOnceWithExactly(secretsStub, {name: 'foobar'});
-        sinon.assert.calledOnce(secretVersionNameStub);
-      });
-
-      it('throws on empty data', async () => {
-        secretsStub = sinon
-          .stub(secretClientStub, 'accessSecretVersion')
-          .resolves([
-            {
-              payload: {
-                data: '',
-              },
-            },
-          ]);
-
-        assert.rejects(bootstrapper.getProbotConfig());
-      });
-
-      it('throws on empty payload', async () => {
-        secretsStub = sinon
-          .stub(secretClientStub, 'accessSecretVersion')
-          .resolves([
-            {
-              payload: {},
-            },
-          ]);
-
-        assert.rejects(bootstrapper.getProbotConfig());
-      });
-
-      it('throws on empty response', async () => {
-        secretsStub = sinon
-          .stub(secretClientStub, 'accessSecretVersion')
-          .resolves([{}]);
-
-        assert.rejects(bootstrapper.getProbotConfig());
-      });
+          },
+        ]);
+      await bootstrapper.getProbotConfig();
+      sinon.assert.calledOnce(secretsStub);
+      sinon.assert.calledOnceWithExactly(secretsStub, {name: 'foobar'});
+      sinon.assert.calledOnce(secretVersionNameStub);
     });
 
-    describe('getLatestSecretVersionName', () => {
-      let bootstrapper: GCFBootstrapper;
-      let secretVersionNameStub: sinon.SinonStub;
+    it('throws on empty data', async () => {
+      secretsStub = sinon
+        .stub(secretClientStub, 'accessSecretVersion')
+        .resolves([
+          {
+            payload: {
+              data: '',
+            },
+          },
+        ]);
 
-      beforeEach(() => {
-        bootstrapper = new GCFBootstrapper();
-        secretVersionNameStub = sinon
-          .stub(bootstrapper, 'getSecretName')
-          .callsFake(() => {
-            return 'foobar';
-          });
-      });
-
-      afterEach(() => {
-        secretVersionNameStub.reset();
-      });
-
-      it('appends "latest"', async () => {
-        const latest = bootstrapper.getLatestSecretVersionName();
-        assert.strictEqual(latest, 'foobar/versions/latest');
-        sinon.assert.calledOnce(secretVersionNameStub);
-      });
+      assert.rejects(bootstrapper.getProbotConfig());
     });
 
-    describe('getSecretName', () => {
-      let bootstrapper: GCFBootstrapper;
-      const storedEnv = process.env;
+    it('throws on empty payload', async () => {
+      secretsStub = sinon
+        .stub(secretClientStub, 'accessSecretVersion')
+        .resolves([
+          {
+            payload: {},
+          },
+        ]);
 
-      beforeEach(() => {
-        bootstrapper = new GCFBootstrapper();
-      });
+      assert.rejects(bootstrapper.getProbotConfig());
+    });
 
-      afterEach(() => {
-        process.env = storedEnv;
-      });
+    it('throws on empty response', async () => {
+      secretsStub = sinon
+        .stub(secretClientStub, 'accessSecretVersion')
+        .resolves([{}]);
 
-      it('formats from env even with nothing', async () => {
-        const latest = bootstrapper.getSecretName();
-        assert.strictEqual(latest, 'projects//secrets/');
-      });
+      assert.rejects(bootstrapper.getProbotConfig());
+    });
+  });
 
-      it('formats from env', async () => {
-        process.env.PROJECT_ID = 'foo';
-        process.env.GCF_SHORT_FUNCTION_NAME = 'bar';
-        const latest = bootstrapper.getSecretName();
-        assert.strictEqual(latest, 'projects/foo/secrets/bar');
-      });
+  describe('getLatestSecretVersionName', () => {
+    let bootstrapper: GCFBootstrapper;
+    let secretVersionNameStub: sinon.SinonStub;
+
+    beforeEach(() => {
+      bootstrapper = new GCFBootstrapper();
+      secretVersionNameStub = sinon
+        .stub(bootstrapper, 'getSecretName')
+        .callsFake(() => {
+          return 'foobar';
+        });
+    });
+
+    afterEach(() => {
+      secretVersionNameStub.reset();
+    });
+
+    it('appends "latest"', async () => {
+      const latest = bootstrapper.getLatestSecretVersionName();
+      assert.strictEqual(latest, 'foobar/versions/latest');
+      sinon.assert.calledOnce(secretVersionNameStub);
+    });
+  });
+
+  describe('getSecretName', () => {
+    let bootstrapper: GCFBootstrapper;
+    const storedEnv = process.env;
+
+    beforeEach(() => {
+      bootstrapper = new GCFBootstrapper();
+    });
+
+    afterEach(() => {
+      process.env = storedEnv;
+    });
+
+    it('formats from env even with nothing', async () => {
+      const latest = bootstrapper.getSecretName();
+      assert.strictEqual(latest, 'projects//secrets/');
+    });
+
+    it('formats from env', async () => {
+      process.env.PROJECT_ID = 'foo';
+      process.env.GCF_SHORT_FUNCTION_NAME = 'bar';
+      const latest = bootstrapper.getSecretName();
+      assert.strictEqual(latest, 'projects/foo/secrets/bar');
     });
   });
 });
